feat(redux): provide returns actions with client
This creates the factories for the `returns` actions, and provides the
respective actions with the client embedded.

BREAKING CHANGE

- The action creators `createReturn`, `fetchReturn`,
`fetchPickupCapabilities`, `fetchReferences`, `fetchReturnsFromOrder`
`updateReturn`, were renamed and now come pre-configured with the
corresponding client from @blackout-client, so you do not need to configure
these action creators anymore if you use the default @blackout-client client provided.

```js
// Previously
import {
  createReturn,
  fetchReturn,
  fetchPickupCapabilities,
  fetchReferences,
  fetchReturnsFromOrder,
  updateReturn,
} from '@blackout-redux/returns';
import {
  postReturn,
  getReturn,
  getPickupCapabilities,
  getReferences,
  getReturnsFromOrder,
  patchReturn,
} from '@blackout-client/returns';

const mapDispatchToProps = {
  createReturn: createReturn(postReturn),
  fetchReturn: fetchReturn(getReturn),
  (...)
};

// Change to:
import {
  createReturn,
  fetchReturn,
  (...)
} from '@blackout-redux/returns';

const mapDispatchToProps = {
  createReturn: createReturn,
  fetchReturn: fetchReturn,
  (...)
};
```
