+++ packages/redux/src/returns/actions/__tests__/fetchReturnsFromOrder.test.js
@@ -1,4 +1,5 @@
 import * as normalizr from 'normalizr';
+import { getReturnsFromOrder } from '@farfetch/blackout-client/returns';
 import { mockStore } from '../../../../tests';
 import {
   responses,
@@ -7,6 +8,11 @@ import {
 import fetchReturnsFromOrder from '../../actions/fetchReturnsFromOrder';
 import find from 'lodash/find';
 import reducer, { actionTypes } from '../../';
+
+jest.mock('@farfetch/blackout-client/returns', () => ({
+  ...jest.requireActual('@farfetch/blackout-client/returns'),
+  getReturnsFromOrder: jest.fn(),
+}));
 
 const returnsMockStore = (state = {}) =>
   mockStore({ returns: reducer() }, state);
@@ -17,8 +23,6 @@ describe('fetchReturnsFromOrder action creator', () => {
   const normalizeSpy = jest.spyOn(normalizr, 'normalize');
   let store;
 
-  const getReturnsFromOrder = jest.fn();
-  const action = fetchReturnsFromOrder(getReturnsFromOrder);
   const orderId = '8VXRHN';
 
   beforeEach(() => {
@@ -33,7 +37,7 @@ describe('fetchReturnsFromOrder action creator', () => {
     expect.assertions(4);
 
     try {
-      await store.dispatch(action(orderId, query));
+      await store.dispatch(fetchReturnsFromOrder(orderId, query));
     } catch (error) {
       expect(error).toBe(expectedError);
       expect(getReturnsFromOrder).toHaveBeenCalledTimes(1);
@@ -58,7 +62,7 @@ describe('fetchReturnsFromOrder action creator', () => {
     getReturnsFromOrder.mockResolvedValueOnce(
       responses.getReturnsFromOrder.get.success,
     );
-    await store.dispatch(action(orderId, query));
+    await store.dispatch(fetchReturnsFromOrder(orderId, query));
 
     const actionResults = store.getActions();
