+++ packages/redux/src/returns/actions/createReturn.js
@@ -1,61 +1,13 @@
-import {
-  CREATE_RETURN_FAILURE,
-  CREATE_RETURN_REQUEST,
-  CREATE_RETURN_SUCCESS,
-} from '../actionTypes';
-import { normalize } from 'normalizr';
-import returnSchema from '../../entities/schemas/return';
+import { createReturnFactory } from './factories';
+import { postReturn } from '@farfetch/blackout-client/returns';
 
 /**
- * @typedef {object} CreateReturnData
- * @property {object} [currentReturn] - Details of the return.
- */
-
-/**
- * @typedef {object} CreateReturnQuery
- * @property {string} [guestUserEmail] - User email. Only required if
- * the user is not registered (guest).
- */
-
-/**
- * @callback CreateReturnThunkFactory
- * @param {CreateReturnData} data  - Details of the Return to be created.
- * @param {CreateReturnQuery} query - Query parameters for creating the return.
- * @param {object} [config] - Custom configurations to send to the client
- * instance (axios).
+ * Create return.
  *
- * @returns {Function} Thunk to be dispatched to the redux store.
- */
-
-/**
- * Method responsible for creating a return.
- *
- * @function createReturn
  * @memberof module:returns/actions
  *
- * @param {Function} postReturn  - Post return client.
+ * @name createReturn
  *
- * @returns {CreateReturnThunkFactory} Thunk factory.
+ * @type {CreateReturnThunkFactory}
  */
-export default postReturn => (data, query, config) => async dispatch => {
-  dispatch({
-    type: CREATE_RETURN_REQUEST,
-  });
-
-  try {
-    const result = await postReturn(data, query, config);
-
-    dispatch({
-      payload: normalize(result, returnSchema),
-      type: CREATE_RETURN_SUCCESS,
-    });
-    return result;
-  } catch (error) {
-    dispatch({
-      payload: { error },
-      type: CREATE_RETURN_FAILURE,
-    });
-
-    throw error;
-  }
-};
+export default createReturnFactory(postReturn);
