+++ packages/redux/src/returns/actions/factories/index.js
@@ -1,0 +1,14 @@
+/**
+ * Returns actions factories.
+ *
+ * @module returns/actions/factories
+ * @category Products
+ * @subcategory Actions Factories
+ */
+
+export { default as createReturnFactory } from './createReturnFactory';
+export { default as fetchPickupCapabilitiesFactory } from './fetchPickupCapabilitiesFactory';
+export { default as fetchReferencesFactory } from './fetchReferencesFactory';
+export { default as fetchReturnFactory } from './fetchReturnFactory';
+export { default as fetchReturnsFromOrderFactory } from './fetchReturnsFromOrderFactory';
+export { default as updateReturnFactory } from './updateReturnFactory';
